feat: Support representing PDB 2.0 debug identifiers (#10)
This adds support for PDB 2.0 debug identifiers which are only a
timestamp and age.  The breakpad format for them is supported as well,
this is the identifier which is supported by microsoft symbol servers
for these identifiers.

The in-memory representation is backwards compatible, so loading
previously serialised objects back will keep working correctly.